refactor: relocate broker auth from runner into core
The broker-auth module (BrokerScope, BrokerAuthStore, extract_bearer_token,
BrokerCredential, ...) lived under `runner/` but depends only on core
(`crate::error` + `crate::paths`) and is shared by the core daemon's broker
endpoints (/exec, /files/*) and the runner layer alike. It is daemon/core auth
infrastructure, not runner behavior.

Move it to `crate::broker_auth` (git mv preserves history). The daemon now uses
`crate::broker_auth::{BrokerScope, extract_bearer_token}` directly instead of
reaching into `crate::runner`, clearing those daemon->runner edges. A
re-export shim in `runner/mod.rs` keeps `crate::runner::{BrokerScope, ...}`
call sites working unchanged; runner-internal module-path callers repoint to
`crate::broker_auth`.

Part of the staged runner extraction. The only remaining daemon->runner prod
edge is `crate::runner::load` in the file-API workspace-root resolver (separate
follow-up); the rest are runner-type imports that leave with the runner crate.
