+++ crates/homeboy-core/src/lib.rs
@@ -94,6 +94,7 @@ pub mod artifact_origin;
 pub mod artifact_postprocess;
 pub mod artifact_preview;
 pub mod artifact_ref;
+pub mod broker_auth;
 pub mod browser_evidence;
 pub mod browser_visual_compare;
 pub mod build_identity;
